made into an ES6 module, included rollup for umd build
